devops: introduce v2 dashboard format
